+++ cheers_lib/src/options.rs
@@ -3,36 +3,36 @@
 #[derive(Clone, Copy)]
 pub struct SearchOptions {
     pub tt_size_mb: usize,
-    pub nmp_depth: i32,
-    pub nmp_reduction: i32,
-    pub see_pruning_depth: i32,
-    pub see_capture_margin: i32,
-    pub see_quiet_margin: i32,
-    pub pvs_fulldepth: i32,
-    pub delta_pruning_margin: i32,
-    pub fp_margin_1: i32,
-    pub fp_margin_2: i32,
-    pub fp_margin_3: i32,
-    pub rfp_margin: i32,
-    pub lmp_depth: i32,
-    pub lmp_margin: i32,
-    pub iir_depth: i32,
+    pub nmp_depth: i8,
+    pub nmp_reduction: i8,
+    pub see_pruning_depth: i8,
+    pub see_capture_margin: i16,
+    pub see_quiet_margin: i16,
+    pub pvs_fulldepth: i8,
+    pub delta_pruning_margin: i16,
+    pub fp_margin_1: i16,
+    pub fp_margin_2: i16,
+    pub fp_margin_3: i16,
+    pub rfp_margin: i16,
+    pub lmp_depth: i8,
+    pub lmp_margin: i16,
+    pub iir_depth: i8,
 }
 
-pub const NMP_DEPTH: i32 = 2;
-pub const NMP_REDUCTION: i32 = 5;
-pub const SEE_PRUNING_DEPTH: i32 = 6;
-pub const SEE_CAPTURE_MARGIN: i32 = 59;
-pub const SEE_QUIET_MARGIN: i32 = 39;
-pub const PVS_FULLDEPTH: i32 = 1;
-pub const DELTA_PRUNING_MARGIN: i32 = 91;
-pub const FP_MARGIN_1: i32 = 115;
-pub const FP_MARGIN_2: i32 = 344;
-pub const FP_MARGIN_3: i32 = 723;
-pub const RFP_MARGIN: i32 = 106;
-pub const LMP_DEPTH: i32 = 1;
-pub const LMP_MARGIN: i32 = 2;
-pub const IIR_DEPTH: i32 = 4;
+pub const NMP_DEPTH: i8 = 2;
+pub const NMP_REDUCTION: i8 = 5;
+pub const SEE_PRUNING_DEPTH: i8 = 6;
+pub const SEE_CAPTURE_MARGIN: i16 = 59;
+pub const SEE_QUIET_MARGIN: i16 = 39;
+pub const PVS_FULLDEPTH: i8 = 1;
+pub const DELTA_PRUNING_MARGIN: i16 = 91;
+pub const FP_MARGIN_1: i16 = 115;
+pub const FP_MARGIN_2: i16 = 344;
+pub const FP_MARGIN_3: i16 = 723;
+pub const RFP_MARGIN: i16 = 106;
+pub const LMP_DEPTH: i8 = 1;
+pub const LMP_MARGIN: i16 = 2;
+pub const IIR_DEPTH: i8 = 4;
 
 impl Default for SearchOptions {
     fn default() -> Self {
